+++ R/ExoLabel.R
@@ -1,7 +1,9 @@
 ExoLabel <- function(edgelistfiles, outfile=tempfile(),
                           mode=c("undirected", "directed"),
+                          weight_type=c("absolute", "probability"),
                           add_self_loops=FALSE,
                           ignore_weights=FALSE,
+                          prob_rescaling=2.0,
                           iterations=0L,
                           return_table=FALSE,
                           consensus_cluster=FALSE,
@@ -9,6 +11,9 @@ ExoLabel <- function(edgelistfiles, outfile=tempfile(),
                           verbose=interactive(),
                           sep='\t',
                           tempfiledir=tempdir()){
+  weight_type <- match.arg(weight_type)
+  if(weight_type == "absolute")
+    prob_rescaling <- 0
   if(!is.numeric(iterations)){
     stop("'iterations' must be an integer or numeric.")
   } else {
@@ -24,6 +29,9 @@ ExoLabel <- function(edgelistfiles, outfile=tempfile(),
   }
   if(!is.numeric(add_self_loops) && !is.logical(add_self_loops)){
     stop("value of 'add_self_loops' should be numeric or logical")
+  }
+  if(!is.numeric(prob_rescaling)){
+    stop("'prob_rescaling' must be numeric")
   }
   if(any(add_self_loops < 0)){
     warning("self loops weight supplied is negative, setting to zero.")
@@ -46,9 +54,17 @@ ExoLabel <- function(edgelistfiles, outfile=tempfile(),
   if(length(add_self_loops) == 1){
     add_self_loops <- rep(add_self_loops, length(outfile))
   }
-  if(length(outfile) != length(add_self_loops)){
-    stop("If more than one outfile is provided, 'add_self_loops' must be ",
-         "either the same length as 'outfile' or length 1.")
+  if(length(prob_rescaling) == 1){
+    prob_rescaling <- rep(prob_rescaling, length(outfile))
+  }
+  if(length(outfile) != length(add_self_loops) ||
+     length(outfile) != length(prob_rescaling)){
+    if(weight_type == "absolute")
+      stop("If more than one outfile is provided, 'add_self_loops' must be ",
+           "either the same length as 'outfile' or length 1.")
+    else
+      stop("If more than one outfile is provided, 'add_self_loops' and 'prob_rescaling' ",
+            "must be either the same length as 'outfile' or length 1.")
   }
   # verify that the first few lines of each file are correct
   if(!all(file.exists(edgelistfiles))) stop("edgelist file does not exist")
@@ -108,26 +124,35 @@ ExoLabel <- function(edgelistfiles, outfile=tempfile(),
 
   if(verbose) cat("Temporary files stored at ", tempfiledir, "\n")
 
+  verbose_int <- 0L
+  if(interactive() && verbose){
+    verbose_int <- 2L
+  } else if(verbose){
+    verbose_int <- 1L
+  }
   seps <- paste(sep, "\n", sep='')
 
   .Call("R_LPOOM_cluster", edgelistfiles, length(edgelistfiles),
         tempfiledir, outfile, seps, iterations,
         verbose, is_undirected, add_self_loops, ignore_weights,
-        consensus_cluster, !use_fast_sort)
+        consensus_cluster, prob_rescaling, !use_fast_sort)
 
   for(f in list.files(tempfiledir, full.names=TRUE))
     if(file.exists(f)) file.remove(f)
   file.remove(tempfiledir)
   retval <- list()
   for(i in seq_along(outfile)){
+    param_vec <- c(add_self_loops=add_self_loops[i], weight_type=weight_type)
+    if(weight_type == "probability")
+      param_vec["prob_rescaling"] <- prob_rescaling[i]
     if(return_table){
       tab <- read.table(outfile[i], sep=sep)
       colnames(tab) <- c("Vertex", "Cluster")
       if(file.exists(outfile[i])) file.remove(outfile[i])
-      retval[[i]] <- list(parameters=c(add_self_loops=add_self_loops[i]),
+      retval[[i]] <- list(parameters=param_vec,
                           results=tab)
     } else {
-      retval[[i]] <- list(parameters=c(add_self_loops=add_self_loops[i]),
+      retval[[i]] <- list(parameters=param_vec,
                           results=outfile[i])
     }
   }