+++ src/OnDiskLP.c
@@ -92,6 +92,9 @@
 #define FILE_READ_CACHE_SIZE 8192*4
 #define CLUSTER_MIN_WEIGHT 0.01
 
+// defines the PID value that maps to 0.5
+#define PID_INFLECTION_VALUE 0.4
+
 /*************/
 /* Constants */
 /*************/
@@ -101,6 +104,12 @@ static const int W_SIZE = sizeof(w_float);
 static const char CONSENSUS_CSRCOPY1[] = "tmpcsr1";
 static const char CONSENSUS_CLUSTER[] = "tmpclust";
 static const int BITS_FOR_WEIGHT = 10;
+
+enum VERBOSITY {
+  VERBOSE_NONE = 0,   // no output
+  VERBOSE_BASIC = 1,  // non-interactive output, better for files
+  VERBOSE_ALL = 2     // interactive output, lots of carriage returns
+};
 
 /*
  * Some comments on external sorting performance:
@@ -180,7 +189,6 @@ typedef struct {
 static l_uint GLOBAL_verts_remaining = 0;
 static leaf **GLOBAL_leaf = NULL;
 static leaf **GLOBAL_all_leaves = NULL;
-static char *GLOBAL_use_node = NULL;
 static char **GLOBAL_filenames = NULL; // holds file NAMES
 static int GLOBAL_nfiles = 0;
 static l_uint CLUST_MAP_CTR = 0;
@@ -193,6 +201,8 @@ static void **GLOBAL_mergebuffers = NULL;
 static LoserTree *GLOBAL_mergetree = NULL;
 static double GLOBAL_max_weight = 1.0;
 static l_uint GLOBAL_verts_changed = 0;
+// c = -1 / log_2(1-x), where x is the value we want to map to 0.5
+static double PID_C = 1;
 
 /***************************/
 /* Struct Helper Functions */
@@ -338,12 +348,21 @@ static void kahan_accu(double *cur_sum, double *cur_err, double new_val){
   return;
 }
 
-inline float sigmoid_transform(const float w, const double slope){
+static inline float sigmoid_transform(const float w, const double slope){
   // should probably expose these at some point
   const float scale = 0.5;
   const float cutoff = 0.1;
   float r = 1 / (1+exp(-1*slope*(w-scale)));
   return r > cutoff ? r : 0;
+}
+
+static inline float pid_transform(const float w, const double shape){
+  // function is \frac{x^{cd}}{x^{cd} + (1-x^c)^d}
+  if(w > 1) error("Attempted to convert probability greater than 1! (received %.2f)", w);
+  double p1 = pow(1-w, PID_C);
+  double p2 = pow(p1, shape), p3 = pow(1-p1, shape);
+  double r = fmax(p2 / (p2+p3), 1.17549435e-38);
+  return -1*log2(r);
 }
 
 static edge compressEdgeValues(l_uint v1, l_uint v2, double weight){
@@ -440,12 +459,6 @@ void cleanup_ondisklp_global_values(){
   if(GLOBAL_queue){
     free_array_queue(GLOBAL_queue);
     GLOBAL_queue = NULL;
-  }
-
-  // free use_node holder
-  if(GLOBAL_use_node){
-    free(GLOBAL_use_node);
-    GLOBAL_use_node = NULL;
   }
 
   // delete any files made during runtime
@@ -598,7 +611,7 @@ static void kway_mergesort_file_inplace(const char* f1, l_uint nlines,
   if(num_bins > (nlines / block_size)){
     num_bins = nlines/block_size + !!(nlines % block_size);
   }
-  if(verbose) Rprintf("\tSorting with %d-way merge...\n", num_bins);
+  if(verbose >= VERBOSE_BASIC) Rprintf("\tSorting with %d-way merge...\n", num_bins);
 
   l_uint nmax_iterations = 1, tmpniter=block_size;
   while(tmpniter*num_bins < nlines){
@@ -626,11 +639,13 @@ static void kway_mergesort_file_inplace(const char* f1, l_uint nlines,
     prev_progress = 0.0;
 
     tmpniter++;
-    if(verbose){
-      Rprintf("\tIteration %" lu_fprint " of %" lu_fprint " (%6.01f%% complete, used ",
+    if(verbose == VERBOSE_ALL){
+      Rprintf("\tIteration %" lu_fprint " of %" lu_fprint " (%5.01f%% complete, used ",
                               tmpniter, nmax_iterations, cur_progress);
       report_filesize(max_fsize);
       Rprintf(")  \r");
+    } else if(verbose == VERBOSE_BASIC){
+      Rprintf("\tIteration %" lu_fprint " of %" lu_fprint "\n", tmpniter, nmax_iterations);
     }
     R_CheckUserInterrupt();
 
@@ -683,12 +698,12 @@ static void kway_mergesort_file_inplace(const char* f1, l_uint nlines,
         cur_progress = ((double)mergetree->nwritten)/nlines*100.0;
         if(cur_progress - prev_progress > 0.5){
           prev_progress = cur_progress;
-          if(verbose){
+          if(verbose == VERBOSE_ALL){
             if(tmpniter == 1){
               cur_fsize = ftell(fileptr);
               if(cur_fsize > max_fsize) max_fsize = cur_fsize;
             }
-            Rprintf("\tIteration %" lu_fprint " of %" lu_fprint " (%6.01f%% complete, used ",
+            Rprintf("\tIteration %" lu_fprint " of %" lu_fprint " (%5.01f%% complete, used ",
                               tmpniter, nmax_iterations, cur_progress);
             report_filesize(max_fsize);
             Rprintf(")  \r");
@@ -709,15 +724,19 @@ static void kway_mergesort_file_inplace(const char* f1, l_uint nlines,
 
       if(cur_progress - prev_progress > 0.5 || cur_progress == 100){
           prev_progress = cur_progress;
-          if(verbose){
+          if(verbose == VERBOSE_ALL){
             if(tmpniter == 1){
               cur_fsize = ftell(fileptr);
               if(cur_fsize > max_fsize) max_fsize = cur_fsize;
             }
-            Rprintf("\tIteration %" lu_fprint " of %" lu_fprint " (%6.01f%% complete, used ",
+            Rprintf("\tIteration %" lu_fprint " of %" lu_fprint " (%5.01f%% complete, used ",
                               tmpniter, nmax_iterations, cur_progress);
             report_filesize(max_fsize);
             Rprintf(")  \r");
+          } else if(verbose == VERBOSE_BASIC && cur_progress == 100){
+            Rprintf("\tIteration complete (used ");
+            report_filesize(ftell(fileptr));
+            Rprintf(")\n");
           }
         R_CheckUserInterrupt();
       }
@@ -729,7 +748,7 @@ static void kway_mergesort_file_inplace(const char* f1, l_uint nlines,
   }
   // cur_source will always be the file we just WROTE to here
 
-  if(verbose) Rprintf("\n");
+  if(verbose >= VERBOSE_BASIC) Rprintf("\n");
   for(int i=0; i<num_bins; i++) free(buffers[i]);
   free(buffers);
   fclose_tracked(1);
@@ -776,7 +795,7 @@ static void kway_mergesort_file(const char* f1, const char* f2, l_uint nlines,
   if(num_bins > (nlines / block_size)){
     num_bins = nlines/block_size + !!(nlines % block_size);
   }
-  if(verbose) Rprintf("\tSorting with %d-way merge...\n", num_bins);
+  if(verbose >= VERBOSE_BASIC) Rprintf("\tSorting with %d-way merge...\n", num_bins);
 
   l_uint nmax_iterations = 1, tmpniter=block_size;
   while(tmpniter*num_bins < nlines){
@@ -804,8 +823,8 @@ static void kway_mergesort_file(const char* f1, const char* f2, l_uint nlines,
     prev_progress = 0.0;
 
     tmpniter++;
-    if(verbose){
-      Rprintf("\tIteration %" lu_fprint " of %" lu_fprint " (%6.01f%% complete, used ",
+    if(verbose == VERBOSE_ALL){
+      Rprintf("\tIteration %" lu_fprint " of %" lu_fprint " (%5.01f%% complete, used ",
                         tmpniter, nmax_iterations, cur_progress);
       report_filesize(max_fsize);
       Rprintf(")  \r");
@@ -861,12 +880,12 @@ static void kway_mergesort_file(const char* f1, const char* f2, l_uint nlines,
         cur_progress = ((double)mergetree->nwritten)/nlines*100.0;
         if(cur_progress - prev_progress > 0.5){
           prev_progress = cur_progress;
-          if(verbose){
+          if(verbose == VERBOSE_ALL){
             if(tmpniter == 1){
               cur_fsize = ftell(file1) + ftell(file2);
               if(cur_fsize > max_fsize) max_fsize = cur_fsize;
             }
-            Rprintf("\tIteration %" lu_fprint " of %" lu_fprint " (%6.01f%% complete, used ",
+            Rprintf("\tIteration %" lu_fprint " of %" lu_fprint " (%5.01f%% complete, used ",
                               tmpniter, nmax_iterations, cur_progress);
             report_filesize(max_fsize);
             Rprintf(")  \r");
@@ -880,15 +899,20 @@ static void kway_mergesort_file(const char* f1, const char* f2, l_uint nlines,
       cur_progress = ((double)mergetree->nwritten)/nlines*100.0;
       if(cur_progress - prev_progress > 0.5 || cur_progress == 100){
         prev_progress = cur_progress;
-        if(verbose){
+        if(verbose == VERBOSE_ALL){
           if(tmpniter == 1){
             cur_fsize = ftell(file1) + ftell(file2);
             if(cur_fsize > max_fsize) max_fsize = cur_fsize;
           }
-          Rprintf("\tIteration %" lu_fprint " of %" lu_fprint " (%6.01f%% complete, used ",
+          Rprintf("\tIteration %" lu_fprint " of %" lu_fprint " (%5.01f%% complete, used ",
                             tmpniter, nmax_iterations, cur_progress);
           report_filesize(max_fsize);
           Rprintf(")  \r");
+        } else if(verbose == VERBOSE_BASIC && cur_progress == 100){
+          Rprintf("\tIteration %" lu_fprint " of %" lu_fprint " (%5.01f%% complete, used ",
+                            tmpniter, nmax_iterations, cur_progress);
+          report_filesize(max_fsize);
+          Rprintf(")  \n");
         }
         R_CheckUserInterrupt();
       }
@@ -906,12 +930,12 @@ static void kway_mergesort_file(const char* f1, const char* f2, l_uint nlines,
   }
   // cur_source will always be the file we just WROTE to here
 
-  if(verbose){
-    Rprintf("\tIteration %" lu_fprint " of %" lu_fprint " (%6.01f%% complete, used ",
+  if(verbose == VERBOSE_ALL){
+    Rprintf("\tIteration %" lu_fprint " of %" lu_fprint " (%5.01f%% complete, used ",
                       tmpniter, nmax_iterations, cur_progress);
     report_filesize(max_fsize);
     Rprintf(")  \r");
-  }
+  } else {}
   R_CheckUserInterrupt();
   for(int i=0; i<num_bins; i++) free(buffers[i]);
   free(buffers);
@@ -991,21 +1015,20 @@ static void split_sorted_file(const char* nfilename, const char* wfilename,
     safe_fread(valuebuffer, L_SIZE, nread, f_r);
 
     // update values
-    for(int i=0; i<nread; i++){
+    for(int i=0; i<nread; i++)
       decompressEdgeValue(valuebuffer[i], &(indices[i]), &(weights[i]));
-    }
 
     safe_fwrite(indices, L_SIZE, nread, f_w);
     safe_fwrite(weights, W_SIZE, nread, wf_w);
     cur_line += nread;
     cur_progress = 100*((double)cur_line) / nlines;
     if(cur_progress - prev_progress > 0.5 || cur_progress == 100){
-      if(v) Rprintf("\tTidying up edges (%6.01f%% Complete)\r", cur_progress);
+      if(v == VERBOSE_ALL) Rprintf("\tTidying up edges (%5.01f%% Complete)\r", cur_progress);
       prev_progress = cur_progress;
       R_CheckUserInterrupt();
     }
   }
-  if(v) Rprintf("\n");
+  if(v == VERBOSE_ALL) Rprintf("\n");
   fclose_tracked(3);
 
   free(weights);
@@ -1107,7 +1130,7 @@ static h_uint hash_file_vnames_trie(const char* fname, prefix *trie, h_uint next
   char c = get_buffchar(read_cache, rc_size, &rcache_i, &remaining, f);
   l_uint print_counter = 0;
 
-  if(v) Rprintf("\tReading file %s...\n", fname);
+  if(v >= VERBOSE_BASIC) Rprintf("\tReading file %s...\n", fname);
 
   while(!feof(f) || remaining){
     // going to assume we're at the beginning of a line
@@ -1161,7 +1184,7 @@ static h_uint hash_file_vnames_trie(const char* fname, prefix *trie, h_uint next
     if(c == line_sep) c = get_buffchar(read_cache, rc_size, &rcache_i, &remaining, f);
     print_counter++;
     if(!(print_counter % PRINT_COUNTER_MOD)){
-      if(v) Rprintf("\t%" lu_fprint " lines read\r", print_counter);
+      if(v == VERBOSE_ALL) Rprintf("\t%" lu_fprint " lines read\r", print_counter);
       else R_CheckUserInterrupt();
     }
   }
@@ -1172,7 +1195,7 @@ static h_uint hash_file_vnames_trie(const char* fname, prefix *trie, h_uint next
         next_index = insert_into_trie(namecache[i], trie, next_index, str_counts[i]);
   }
 
-  if(v) Rprintf("\t%" lu_fprint " lines read\n", print_counter);
+  if(v >= VERBOSE_BASIC) Rprintf("\t%" lu_fprint " lines read\n", print_counter);
   fclose_tracked(1);
   GLOBAL_max_weight = max_weight;
 
@@ -1209,7 +1232,7 @@ static l_uint reindex_trie_and_write_counts(prefix *trie, FILE* csrfile,
     (*nseen)++;
     if(((*nseen)+1) % PRINT_COUNTER_MOD == 0){
       R_CheckUserInterrupt();
-      if(verbose) Rprintf("\tProcessed %" lu_fprint " vertices\r", *nseen);
+      if(verbose == VERBOSE_ALL) Rprintf("\tProcessed %" lu_fprint " vertices\r", *nseen);
     }
   }
   while(ctr < bits_remaining)
@@ -1226,22 +1249,6 @@ static void reset_trie_clusters(l_uint num_v){
   for(l_uint i=0; i<num_v; i++){
     l = GLOBAL_all_leaves[i];
     l->count = l->index+1;
-  }
-
-  return;
-}
-
-static void partial_reset_clusters(l_uint num_v, l_uint clust_val){
-  leaf *l;
-  for(l_uint i=0; i<num_v; i++){
-    l = GLOBAL_all_leaves[i];
-    if(clust_val+1 == l->count){ // 1-indexed clusters
-      l->count = l->index+1;
-      GLOBAL_use_node[l->index] = 1;
-
-    } else {
-      GLOBAL_use_node[l->index] = 0;
-    }
   }
 
   return;
@@ -1329,7 +1336,7 @@ static l_uint csr_compress_edgelist_trie_batch(const char* edgefile, prefix *tri
   // note: weights/neighbors don't need to be initialized
   // see https://stackoverflow.com/questions/31642389/fseek-a-newly-created-file
 
-  if(v) Rprintf("\tReading file %s...\n", edgefile);
+  if(v >= VERBOSE_BASIC) Rprintf("\tReading file %s...\n", edgefile);
 
   char c = get_buffchar(read_cache, rc_size, &rcache_i, &remaining, edgelist);
   while(!feof(edgelist) || remaining){ // file isn't done OR buffer has data
@@ -1392,12 +1399,12 @@ static l_uint csr_compress_edgelist_trie_batch(const char* edgefile, prefix *tri
 
     print_counter++;
     if(!(print_counter % PRINT_COUNTER_MOD)){
-      if(v) Rprintf("\t%" lu_fprint " edges read\r", print_counter);
+      if(v == VERBOSE_ALL) Rprintf("\t%" lu_fprint " edges read\r", print_counter);
       R_CheckUserInterrupt();
     }
   }
 
-  if(v) Rprintf("\t%" lu_fprint " edges read\n", print_counter);
+  if(v >= VERBOSE_BASIC) Rprintf("\t%" lu_fprint " edges read\n", print_counter);
   fclose_tracked(3);
   free(edges);
   free(vname);
@@ -1417,7 +1424,6 @@ static void add_remaining_to_queue(l_uint new_clust, leaf **neighbors,
     tmp_cl = neighbors[i]->count;
     if(tmp_cl == new_clust || weights[i] < CLUSTER_MIN_WEIGHT) continue;
     tmp_ind = neighbors[i]->index;
-    if(!GLOBAL_use_node[tmp_ind]) continue;
     found = 0;
     for(int j=0; j<ctr; j++){
       if(neighbors[j]->index == tmp_ind){
@@ -1437,7 +1443,8 @@ static void add_remaining_to_queue(l_uint new_clust, leaf **neighbors,
   return;
 }
 
-static void update_node_cluster(l_uint ind, float self_loop_weight,
+static void update_node_cluster(l_uint ind,
+                          float self_loop_weight, float pid_shape,
                           aq_int times_seen,
                           FILE *offsets, FILE *weightsfile, FILE *neighborfile,
                           ArrayQueue *queue, float atten_param){
@@ -1464,6 +1471,8 @@ static void update_node_cluster(l_uint ind, float self_loop_weight,
   l_uint *clusts;
   l_uint *indices;
   leaf **neighbors;
+
+  if(pid_shape != 0) self_loop_weight = pid_transform(self_loop_weight, pid_shape);
 
   // move to information for the vertex and read in number of edges
   // TODO: can we put this in the trie?
@@ -1493,7 +1502,7 @@ static void update_node_cluster(l_uint ind, float self_loop_weight,
   // read in the clusters
   for(l_uint i=0; i<num_edges; i++){
     neighbors[i] = GLOBAL_all_leaves[clusts[i]];
-
+    if(pid_shape != 0) weights_arr[i] = pid_transform(weights_arr[i], pid_shape);
     // attenuate edges, using adaptive scaling
     // see https://doi.org/10.1103/PhysRevE.83.036103, eqns 4-5
     weights_arr[i] *= 1-(atten_param * neighbors[i]->dist);
@@ -1574,7 +1583,7 @@ static void update_node_cluster(l_uint ind, float self_loop_weight,
 static void cluster_file(const char* offsets_fname, const char* weights_fname,
                           const char* neighbor_fname,
                           l_uint num_v, int max_iterations, int v,
-                          float self_loop_weight){
+                          float self_loop_weight, float pid_shape){
   GLOBAL_verts_remaining = num_v;
 
   // main runner function to cluster nodes
@@ -1601,12 +1610,12 @@ static void cluster_file(const char* offsets_fname, const char* weights_fname,
 
   // randomly initialize queue and ctr file
 
-  if(v) Rprintf("\tInitializing queues...");
+  if(v >= VERBOSE_BASIC) Rprintf("\tInitializing queues...");
   GLOBAL_queue = init_array_queue(num_v, max_iterations);
   iteration_length = GLOBAL_queue->length;
-  if(v) Rprintf("done.\n\tClustering network:\n");
+  if(v >= VERBOSE_BASIC) Rprintf("done.\n\tClustering network:\n");
 
-  if(v) Rprintf("\t0.0%% complete %s", progress[++statusctr%progbarlen]);
+  if(v == VERBOSE_ALL) Rprintf("\t0.0%% complete %s", progress[++statusctr%progbarlen]);
   /*
    * TODO: we can parallelize this, right?
    *    - set up the number of threads
@@ -1624,7 +1633,7 @@ static void cluster_file(const char* offsets_fname, const char* weights_fname,
     }
     print_counter++;
     if(!(print_counter % PROGRESS_COUNTER_MOD)){
-      if(v){
+      if(v == VERBOSE_ALL){
         pct_complete = ((float)(num_v - GLOBAL_queue->length)) / num_v;
         if(pct_complete != 1 && fabs(pct_complete - prev_pct) < MIN_UPDATE_PCT){
           pct_complete = prev_pct;
@@ -1638,20 +1647,21 @@ static void cluster_file(const char* offsets_fname, const char* weights_fname,
     }
     l_uint next_vert = array_queue_pop(GLOBAL_queue);
     iteration_length--;
-    if(!GLOBAL_use_node[next_vert]) continue; // skip nodes not being evaluated
+
     // will either be negative (because just removed from queue)
     // or zero (because seen max_iteration times)
     aq_int num_seen = -1*GLOBAL_queue->seen[next_vert];
     if(!num_seen) num_seen = max_iterations;
-    update_node_cluster(next_vert, self_loop_weight, num_seen,
+    update_node_cluster(next_vert, self_loop_weight, pid_shape, num_seen,
                         offsetsfile, weightsfile, neighborfile,
                         GLOBAL_queue, atten_param);
   }
-  if(v){
+  if(v >= VERBOSE_BASIC){
+    if(v == VERBOSE_ALL) Rprintf("\r");
     if(max_iterations > 0)
-      Rprintf("\r\t100%% complete!                \n");
+      Rprintf("\t100%% complete!                \n");
     else
-      Rprintf("\r\tComplete! (%d total iterations)     \n", i+1);
+      Rprintf("\tComplete! (%d total iterations)     \n", i+1);
   }
   free_array_queue(GLOBAL_queue);
   GLOBAL_queue = NULL;
@@ -1790,7 +1800,7 @@ static void resolve_cluster_consensus(const char* clusters, const char* csrheade
 static void consensus_cluster_oom(const char* csrfile, const char* weightsfile,
                             const char* neighborfile, const char* dir,
                             l_uint num_v, int num_iter, int v,
-                            float self_loop_weight,
+                            float self_loop_weight, float pid_shape,
                             const double* consensus_weights, const int consensus_len){
 
   /*
@@ -1824,10 +1834,10 @@ static void consensus_cluster_oom(const char* csrfile, const char* weightsfile,
   dummyclust = safe_fopen(tmpclusterfile, "wb");
   // now we run clustering over consensus_len times
   for(int i=0; i<consensus_len; i++){
-    if(v) Rprintf("Iteration %d of %d:\n", i+1, consensus_len);
+    if(v >= VERBOSE_BASIC) Rprintf("Iteration %d of %d:\n", i+1, consensus_len);
 
     // modify weights according to sigmoid transformation
-    if(v) Rprintf("\tTransforming edge weights...\n");
+    if(v >= VERBOSE_BASIC) Rprintf("\tTransforming edge weights...\n");
     copy_weightsfile_sig(transformedweights, weightsfile, num_edges, consensus_weights[i]);
 
     // reset cluster values
@@ -1835,9 +1845,9 @@ static void consensus_cluster_oom(const char* csrfile, const char* weightsfile,
 
     // cluster with transformed weights
     cluster_file(csrfile, transformedweights, neighborfile,
-                  num_v, num_iter, v, self_loop_weight);
+                  num_v, num_iter, v, self_loop_weight, pid_shape);
 
-    if(v) Rprintf("\tRecording results...\n");
+    if(v >= VERBOSE_BASIC) Rprintf("\tRecording results...\n");
     for(l_uint i=0; i<num_v; i++){
       tmpleaf = GLOBAL_all_leaves[i];
       clusters[tmpleaf->index] = tmpleaf->count;
@@ -1851,13 +1861,13 @@ static void consensus_cluster_oom(const char* csrfile, const char* weightsfile,
   free(clusters);
 
   // now we can just destroy the other files
-  if(v) Rprintf("Reconciling runs...\n");
+  if(v >= VERBOSE_BASIC) Rprintf("Reconciling runs...\n");
   resolve_cluster_consensus(tmpclusterfile, csrfile, weightsfile, neighborfile,
                             num_v, consensus_len);
 
-  if(v) Rprintf("Clustering on consensus data...\n");
+  if(v >= VERBOSE_BASIC) Rprintf("Clustering on consensus data...\n");
   reset_trie_clusters(num_v);
-  cluster_file(csrfile, weightsfile, neighborfile, num_v, num_iter, v, self_loop_weight);
+  cluster_file(csrfile, weightsfile, neighborfile, num_v, num_iter, v, self_loop_weight, pid_shape);
 
   return;
 }
@@ -1894,7 +1904,7 @@ static l_uint write_output_clusters_trie(FILE *outfile, prefix *trie, l_uint *cl
         safe_fwrite(write_buf, 1, strlen(write_buf), outfile);
         num_v--;
         if(num_v % PROGRESS_COUNTER_MOD == 0){
-          if(verbose)
+          if(verbose == VERBOSE_ALL)
             Rprintf("\r\tNodes remaining: %" lu_fprint "                    ", num_v);
           else
             R_CheckUserInterrupt();
@@ -1933,8 +1943,8 @@ SEXP R_LPOOM_cluster(SEXP FILENAME, SEXP NUM_EFILES, // files
                     SEXP OUTDIR, SEXP OUTFILES,  // more files
                     SEXP SEPS, SEXP ITER, SEXP VERBOSE, // control flow
                     SEXP IS_UNDIRECTED, SEXP ADD_SELF_LOOPS, // optional adjustments
-                    SEXP IGNORE_WEIGHTS,SEXP CONSENSUS_WEIGHTS,
-                    SEXP SORT_INPLACE){
+                    SEXP IGNORE_WEIGHTS, SEXP CONSENSUS_WEIGHTS,
+                    SEXP PID_SHAPE, SEXP SORT_INPLACE){
   /*
    * I always forget how to handle R strings so I'm going to record it here
    * R character vectors are STRSXPs, which is the same as a list (VECSXP)
@@ -1963,7 +1973,7 @@ SEXP R_LPOOM_cluster(SEXP FILENAME, SEXP NUM_EFILES, // files
   GLOBAL_nbuffers = 0;
   GLOBAL_mergetree = NULL;
   GLOBAL_max_weight = 1.0;
-  GLOBAL_use_node = NULL;
+  PID_C = -1/log2(1-PID_INFLECTION_VALUE);
 
   // main files
   const char* dir = CHAR(STRING_ELT(OUTDIR, 0));
@@ -1979,12 +1989,13 @@ SEXP R_LPOOM_cluster(SEXP FILENAME, SEXP NUM_EFILES, // files
   const char* seps = CHAR(STRING_ELT(SEPS, 0));
   const int num_edgefiles = INTEGER(NUM_EFILES)[0];
   aq_int num_iter = INTEGER(ITER)[0];
-  const int verbose = LOGICAL(VERBOSE)[0];
+  const int verbose = INTEGER(VERBOSE)[0];
   l_uint num_v = 0;
 
   // optional parameters
   const int is_undirected = LOGICAL(IS_UNDIRECTED)[0];
   const double* self_loop_weights = REAL(ADD_SELF_LOOPS);
+  const double* pid_shape = REAL(PID_SHAPE);
   const int ignore_weights = LOGICAL(IGNORE_WEIGHTS)[0];
   const int use_inplace_sort = LOGICAL(SORT_INPLACE)[0];
 
@@ -1997,20 +2008,20 @@ SEXP R_LPOOM_cluster(SEXP FILENAME, SEXP NUM_EFILES, // files
 
   // first hash all vertex names
   time1 = clock();
-  if(verbose) Rprintf("Building trie for vertex names...\n");
+  if(verbose >= VERBOSE_BASIC) Rprintf("Building trie for vertex names...\n");
   for(int i=0; i<num_edgefiles; i++){
     edgefile = CHAR(STRING_ELT(FILENAME, i));
     num_v = hash_file_vnames_trie(edgefile, GLOBAL_trie, num_v, seps[0], seps[1],
                                   verbose, is_undirected, ignore_weights);
   }
   time2 = clock();
-  if(verbose) report_time(time1, time2, "\t");
+  if(verbose >= VERBOSE_BASIC) report_time(time1, time2, "\t");
 
   // allocate space for leaf counters
   GLOBAL_all_leaves = safe_malloc(sizeof(leaf *) * (num_v+1));
 
   // next, reformat the file to get final counts for each node
-  if(verbose) Rprintf("Tidying up internal tables...\n");
+  if(verbose >= VERBOSE_BASIC) Rprintf("Tidying up internal tables...\n");
   //num_v = node_vertex_file_cleanup(dir, hashfile, temptabfile, hashindex, verbose);
   FILE *tempcounts = safe_fopen(temptabfile, "wb");
   if(!tempcounts) error("could not open file %s\n", temptabfile);
@@ -2019,7 +2030,7 @@ SEXP R_LPOOM_cluster(SEXP FILENAME, SEXP NUM_EFILES, // files
   l_uint max_degree = reindex_trie_and_write_counts(GLOBAL_trie, tempcounts, 0, verbose, &print_val);
   fclose_tracked(1);
 
-  if(verbose) Rprintf("\tFound %" lu_fprint " unique vertices!\n", num_v);
+  if(verbose >= VERBOSE_BASIC) Rprintf("\tFound %" lu_fprint " unique vertices!\n", num_v);
   if(!num_iter){
     // + 2 is to add an extra in case there are self loops
     // may have an extra, but like at the end of the day, it's one extra degree
@@ -2029,16 +2040,16 @@ SEXP R_LPOOM_cluster(SEXP FILENAME, SEXP NUM_EFILES, // files
     size_t num_bits = sizeof(aq_int) * 8 - 1; // signed, so we have one less bit to work with
     num_iter = ((aq_int)(1)) << num_bits;
     num_iter = num_iter > max_degree ? max_degree : num_iter;
-    if(verbose) Rprintf("\tAutomatically setting iterations to %d\n", num_iter);
+    if(verbose >= VERBOSE_BASIC) Rprintf("\tAutomatically setting iterations to %d\n", num_iter);
   }
 
   // next, create an indexed table file of where data for each vertex will be located
-  if(verbose) Rprintf("Reformatting vertex degree file...\n");
+  if(verbose >= VERBOSE_BASIC) Rprintf("Reformatting vertex degree file...\n");
   reformat_counts(temptabfile, tabfile, num_v);
 
   // then, we'll create the CSR compression of all our edges
   time1 = clock();
-  if(verbose) Rprintf("Reading in edges...\n");
+  if(verbose >= VERBOSE_BASIC) Rprintf("Reading in edges...\n");
   l_uint num_edges = 0;
   for(int i=0; i<num_edgefiles; i++){
     edgefile = CHAR(STRING_ELT(FILENAME, i));
@@ -2072,58 +2083,28 @@ SEXP R_LPOOM_cluster(SEXP FILENAME, SEXP NUM_EFILES, // files
   time2 = clock();
   //check_mergedsplit_file(neighborfile, weightsfile);
 
-  if(verbose) report_time(time1, time2, "\t");
-
-  // use all nodes at this stage
-  GLOBAL_use_node = safe_malloc(num_v);
+  if(verbose >= VERBOSE_BASIC) report_time(time1, time2, "\t");
 
   char vert_name_holder[MAX_NODE_NAME_SIZE];
   char write_buffer[PATH_MAX];
   for(int i=0; i<num_ofiles; i++){
-    memset(GLOBAL_use_node, 1, num_v);
     reset_trie_clusters(num_v);
     time1 = clock();
+    // set the c value to self_loop_weight, or 0.5 if none provided
+    PID_C = self_loop_weights[i] != 0 ? self_loop_weights[i] : 0.5;
+    PID_C = -1/log2(1-PID_C);
     if(consensus_len){
       consensus_cluster_oom(tabfile, weightsfile, neighborfile, dir, num_v,
-                            num_iter, verbose, self_loop_weights[i],
+                            num_iter, verbose, self_loop_weights[i], pid_shape[i],
                             consensus_w, consensus_len);
 
     } else {
-      if(verbose) Rprintf("Clustering...\n");
+      if(verbose >= VERBOSE_BASIC) Rprintf("Clustering...\n");
       cluster_file(tabfile, weightsfile, neighborfile, num_v, num_iter, verbose,
-                    self_loop_weights[i]);
+                    self_loop_weights[i], pid_shape[i]);
     }
     time2 = clock();
-    if(verbose) report_time(time1, time2, "\t");
-
-    // re-cluster all the clusters
-    l_uint *cluster_counts = safe_calloc(num_v, L_SIZE);
-    // note clusters are 1-indexed, so have to subtract 1
-    for(l_uint i=0; i<num_v; i++)
-      cluster_counts[GLOBAL_all_leaves[i]->count - 1]++;
-
-    /*
-    // re-cluster all clusters
-    time1 = clock();
-    if(verbose) Rprintf("Reclustering...");
-    double pct_complete = 0, last_pct = 0;
-    for(l_uint i=0; i<num_v; i++){
-      if(cluster_counts[i] > 2){
-        partial_reset_clusters(num_v, i);
-        // run with verbose = FALSE so we don't get a bajillion outputs
-        cluster_file(tabfile, weightsfile, neighborfile, num_v, num_iter, 0,
-                      inflations[i], self_loop_weights[i]);
-      }
-      pct_complete = ((double)(i+1) / num_v) * 100;
-      if(pct_complete == 100 || pct_complete > last_pct+0.5){
-        last_pct = pct_complete;
-        if(verbose) Rprintf("Reclustering...(%5.1f%% complete)\r", pct_complete);
-        else R_CheckUserInterrupt();
-      }
-    }
-    time2 = clock();
-    if(verbose) report_time(time1, time2, "\t");
-    */
+    if(verbose >= VERBOSE_BASIC) report_time(time1, time2, "\t");
 
     // have to allocate resources for writing out
     outfile = CHAR(STRING_ELT(OUTFILES, i));
@@ -2131,10 +2112,12 @@ SEXP R_LPOOM_cluster(SEXP FILENAME, SEXP NUM_EFILES, // files
     if(!results) error("Failed to open output file.");
     l_uint *clust_mapping = safe_calloc(num_v, L_SIZE);
     CLUST_MAP_CTR = 1;
-    if(verbose) Rprintf("Writing clusters to file...\n\tVertices remaining: %" lu_fprint "", num_v);
+    if(verbose == VERBOSE_ALL) Rprintf("Writing clusters to file...\n\tVertices remaining: %" lu_fprint "", num_v);
+    else if(verbose == VERBOSE_BASIC) Rprintf("Writing clusters to file...\n");
     write_output_clusters_trie(results, GLOBAL_trie, clust_mapping, vert_name_holder,
                                 0, write_buffer, (size_t)PATH_MAX, seps, num_v, verbose);
-    if(verbose) Rprintf("\r\tNodes remaining: Done!               \n");
+    if(verbose == VERBOSE_ALL) Rprintf("\r\tNodes remaining: Done!               \n");
+    else if(verbose == VERBOSE_BASIC) Rprintf("Execution completed!\n");
     free(clust_mapping);
     fclose_tracked(1);
   }