+++ NEWS.md
@@ -1,6 +1,8 @@
 # SynExtend 1.19.7
 * `ExoLabel` now uses hop-length attenuation to mitigate formation of massive communities.
 * `ExoLabel` no longer supports `inflation`, since attentuation does a better job handling this without introducing additional parameters.
+* Adds probablistic edge support for `ExoLabel`
+* `ExoLabel` will now print a lot less when running with `verbose=TRUE` in non-interactive mode -- just as informative, but less junk caused by lots of unrendered carriage returns.
 
 # SynExtend 1.19.6
 * fixes multiple bugs in EstimRearrScen